- 공연 조회 컨트럴로를 임시로 reservation 모듈로 위치 변경 - url규칙을 정함에 따라, 공연 상세 조회 URL을 변경 - DTO 이름을 구체적으로 변경
